remove duplicate gaussian code and restore tomo utils
