+++ src/quantem/tomography/utils.py
@@ -127,44 +127,8 @@ def get_TV_loss(tensor, factor=1e-3):
 
     return tv_loss * factor / (torch.prod(torch.tensor(tensor.shape)))
 
-
-# --- Gaussian filters ---
-
-
-def gaussian_kernel_1d(sigma: float, num_sigmas: float = 3.0) -> torch.Tensor:
-    radius = np.ceil(num_sigmas * sigma)
-    support = torch.arange(-radius, radius + 1, dtype=torch.float)
-    kernel = torch.distributions.Normal(loc=0, scale=sigma).log_prob(support).exp_()
-    # Ensure kernel weights sum to 1, so that image brightness is not altered
-    return kernel.mul_(1 / kernel.sum())
-
-
-def gaussian_filter_2d(
-    img: torch.Tensor, sigma: float, kernel_1d: torch.Tensor
-) -> torch.Tensor:  # Add kernel_1d as an argument
-    # kernel_1d = gaussian_kernel_1d(sigma)  # Create 1D Gaussian kernel - Moved outside function
-    padding = len(kernel_1d) // 2  # Ensure that image size does not change
-    img = img.unsqueeze(0).unsqueeze_(0)  # Make copy, make 4D for ``conv2d()``
-    # Convolve along columns and rows
-    img = torch.nn.functional.conv2d(img, weight=kernel_1d.view(1, 1, -1, 1), padding=(padding, 0))
-    img = torch.nn.functional.conv2d(img, weight=kernel_1d.view(1, 1, 1, -1), padding=(0, padding))
-    return img.squeeze_(0).squeeze_(0)  # Make 2D again
-
-
-def gaussian_filter_1d(
-    arr: torch.Tensor, kernel_1d: torch.Tensor
-) -> torch.Tensor:  # Replicate-padded torch alternative to ``scipy.ndimage.gaussian_filter1d``
-    padding = len(kernel_1d) // 2  # Ensure that signal size does not change
-    arr = arr.unsqueeze(0).unsqueeze_(0)  # Make copy, make 3D for ``conv1d()``
-    # Replicate edge values so the output has no zero-padded ringing at boundaries
-    arr = torch.nn.functional.pad(arr, (padding, padding), mode="replicate")
-    arr = torch.nn.functional.conv1d(arr, weight=kernel_1d.view(1, 1, -1))
-    return arr.squeeze_(0).squeeze_(0)  # Make 1D again
-
-
-def gaussian_filter_2d_stack(stack: torch.Tensor, kernel_1d: torch.Tensor) -> torch.Tensor:
-    """
-    Apply 2D Gaussian blur to each slice stack[:, i, :] in a vectorized way.
+    Encourages piecewise smoothness by penalizing differences between
+    adjacent elements.
 
     Args:
         x:         Input tensor of shape (N, C, L) or (N, L) or (L,)
